Move _find_closest to _helpers
