+++ plantcv/plantcv/hyperspectral/__init__.py
@@ -1,4 +1,3 @@
-from plantcv.plantcv.hyperspectral.read_data import _find_closest
 from plantcv.plantcv.hyperspectral.read_data import _make_pseudo_rgb
 from plantcv.plantcv.hyperspectral.read_data import read_data
 from plantcv.plantcv.hyperspectral.extract_wavelength import extract_wavelength
@@ -9,6 +8,5 @@ from plantcv.plantcv.hyperspectral.rot90 import rot90
 from plantcv.plantcv.hyperspectral.write_data import write_data
 
 # add new functions to end of lists
-__all__ = ["read_data", "_find_closest", "calibrate",
            "_make_pseudo_rgb", "extract_wavelength", "_avg_reflectance", "_inverse_covariance",
            "rot90", "write_data"]
