+++ plantcv/plantcv/hyperspectral/extract_wavelength.py
@@ -5,7 +5,7 @@ import numpy as np
 from plantcv.plantcv import params
 from plantcv.plantcv._debug import _debug
 from plantcv.plantcv import Spectral_data
-from plantcv.plantcv.hyperspectral.read_data import _find_closest
+from plantcv.plantcv._helpers import _find_closest
 
 
 def extract_wavelength(spectral_data, wavelength):
